requirements.txt
numpy>=1.19.2
pandas>=1.2.3
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.2.0
scikit-learn>=0.24.1
matplotlib>=3.4.1
opencv-python>=4.5.1
tqdm>=4.60.0